feat: you-goal-p10-wire-decision-routing-002 - Route structured reviewer envelopes from the parsed decision field
